+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 1.x branch
 ## 1.7 branch
 ### 1.7.2-RC2
-* `InstagramTrait::user()` method now returns an `Entity` instance;
+* `InstagramTrait::user()` and `InstagramTrait::media()` methods now return
+    `Entity` instances;
 * `InstagramTrait::photos()` method now returns an array with entities of photos;
 * updated for MeCms 2.22.5-RC2.